+++ source/MaterialXGraphEditor/Main.cpp
@@ -5,9 +5,10 @@
 
 #include <MaterialXGraphEditor/Graph.h>
 
-#include <GLFW/glfw3.h>
-
+#include "imgui_impl_glfw.h"
 #include "imgui_impl_opengl3.h"
+
+#include <GLFW/glfw3.h>
 
 #include <iostream>
